一个可以正常编译执行的版本，下一版本将开始改动mbr ------- 修复了之前存在的bug：由于现在实模式段是0x7c0，所以清空prefix区应当切换到0x0段使用 引入了VBRmain.cpp， 原来main.cpp被改成了MBRmain.cpp，含义更加明确 MBRmain.cpp生成的函数尺寸被优化，不超过分区表的开始

+++ include/PMLoader.h
@@ -155,8 +155,8 @@ public:
     *at paddr: .word(length) .long(baseaddr).  baseaddr在高地址
     *相对于栈来说，较后面的参数在栈高地址
     */
-    static void setidtr(short len,int address);
-    static void setgdtr(short len,int address);
+    static void setidtr(int len,int address);
+    static void setgdtr(int len,int address);
     /**
     * 必须提供PROTECTED_SEL,PROTECTED_ENTRY的值
     * 或者必须将保护模式代码置于 0b10000:0 处开始
